staticReports in almost done
