+++ frontend/src/components/landing/LandingPage.tsx
@@ -3,7 +3,7 @@ import type { Report } from "@/shared/types/db-models";
 import { getOurReports } from "../../services/ourReports";
 import { Header } from "../general/header/Header";
 import { Footer } from "../general/footer/Footer";
-import { StaticReportComp } from "./staticReports/StaticReports";
+import { StaticReports } from "./staticReports/StaticReports";
 
 import './landingPage.scss'
 
@@ -21,30 +21,7 @@ export function LandingPage() {
   return (
     <>
       <Header />
-      <StaticReportComp 
-        id={1} 
-        title="Título" 
-        subtitle="Subtitulo" 
-        image={'staticReports/FotoExample.jpg'}
-        paragraph="Texto textito textado texteado textando 
-        textaso texting textis text.  textetextuando textaxado 
-        textis, tuxto oxtus"/>
-      <StaticReportComp 
-        id={1} 
-        title="Título" 
-        subtitle="Subtitulo" 
-        image={'staticReports/FotoExample.jpg'}
-        paragraph="Texto textito textado texteado textando 
-        textaso texting textis text.  textetextuando textaxado 
-        textis, tuxto oxtus"/>
-      <StaticReportComp 
-        id={1} 
-        title="Título" 
-        subtitle="Subtitulo" 
-        image={'staticReports/FotoExample.jpg'}
-        paragraph="Texto textito textado texteado textando 
-        textaso texting textis text.  textetextuando textaxado 
-        textis, tuxto oxtus"/>
+      <StaticReports/>
       <Footer />
     </>
   );
