+++ frontend/src/components/landing/staticReports/individualStaticReport/staticReportComp.scss
@@ -1,4 +1,4 @@
-@use '../../../Global' as *;
+@use '../../../../Global' as *;
 
 div.static-report {
   padding: 0 20px;
